Consolidate compile.Compiler and utils.compiler
