+++ setup.py
@@ -54,13 +54,13 @@ packages = \
  'starknet_py.net.l1',
  'starknet_py.net.models',
  'starknet_py.utils',
- 'starknet_py.utils.compiler',
  'starknet_py.utils.crypto',
  'starknet_py.utils.data_transformer',
  'starknet_py.utils.sync']
 
 package_data = \
-{'': ['*']}
+{'': ['*'],
+ 'starknet_py.compile': ['mock-contracts/*', 'mock-contracts/inner/*']}
 
 install_requires = \
 ['asgiref>=3.4.1,<4.0.0', 'cairo-lang==0.7.1', 'web3>=5.28.0,<6']
