+++ starknet_py/compile/compiler.py
@@ -1,8 +1,17 @@
-from typing import List, NewType, Optional, Union
+import json
+import os
+from pathlib import Path
+from typing import List, NewType, Optional, Tuple, Union
 
 from starkware.starknet.services.api.contract_definition import ContractDefinition
-
-from starknet_py.utils.compiler.starknet_compile import starknet_compile
+from starkware.cairo.lang.compiler.constants import MAIN_SCOPE, LIBS_DIR_ENVVAR
+from starkware.cairo.lang.cairo_constants import DEFAULT_PRIME
+from starkware.cairo.lang.compiler.cairo_compile import (
+    get_module_reader,
+)
+from starkware.cairo.lang.compiler.preprocessor.preprocess_codes import preprocess_codes
+from starkware.starknet.compiler.compile import assemble_starknet_contract
+from starkware.starknet.compiler.starknet_pass_manager import starknet_pass_manager
 
 CairoSourceCode = NewType("CairoSourceCode", str)
 CairoFilename = NewType("CairoFilename", str)
@@ -16,9 +25,6 @@ class Compiler:
     Class for compiling Cairo contracts
     """
 
-    def __init__(self):
-        pass
-
     @staticmethod
     def create_contract_definition(
         compiled_contract: Optional[str] = None,
@@ -28,10 +34,11 @@ class Compiler:
         """
         Creates ContractDefinition either from already compiled contract or contract source code
 
+        :raises ValueError: when neither contract_source and compiled_contract are given
         :param compiled_contract: an already compiled contract
         :param contract_source: string containing source code or a list of source files paths
         :param search_paths: a ``list`` of paths used by starknet_compile to resolve dependencies within contracts
-        :return a ContractDefinition
+        :return: a ContractDefinition
         """
         if not contract_source and not compiled_contract:
             raise ValueError(
@@ -55,6 +62,66 @@ class Compiler:
 
         :param compilation_source: string containing source code or a list of source files paths
         :param search_paths: a ``list`` of paths used by starknet_compile to resolve dependencies within contracts
-        :return string of compiled contract
+        :return: string of compiled contract
         """
-        return starknet_compile(compilation_source, search_paths=search_paths)
+        return Compiler._compile(compilation_source, search_paths=search_paths)
+
+    @staticmethod
+    def _load_cairo_source_code(filename: CairoFilename) -> str:
+        source_file = Path(filename)
+
+        if not source_file.is_file():
+            raise TypeError(f"{filename} does not exist")
+
+        if source_file.suffix != ".cairo":
+            raise TypeError(f"{filename} is not a cairo source file")
+
+        return Path(filename).read_text("utf-8")
+
+    @staticmethod
+    def _load_source_code(src: StarknetCompilationSource) -> List[Tuple[str, str]]:
+        if isinstance(src, str):
+            return [(src, str(hash(src)))]
+        return [
+            (Compiler._load_cairo_source_code(filename), filename) for filename in src
+        ]
+
+    @staticmethod
+    def _compile(
+        source: StarknetCompilationSource, search_paths: Optional[List[str]] = None
+    ):
+        file_contents_for_debug_info = {}
+
+        cairo_path: List[str] = list(
+            filter(None, os.getenv(LIBS_DIR_ENVVAR, "").split(":"))
+        )
+
+        if search_paths is not None:
+            cairo_path += search_paths
+
+        module_reader = get_module_reader(cairo_path=cairo_path)
+
+        pass_manager = starknet_pass_manager(
+            prime=DEFAULT_PRIME,
+            read_module=module_reader.read,
+            disable_hint_validation=True,
+        )
+
+        preprocessed = preprocess_codes(
+            codes=Compiler._load_source_code(source),
+            pass_manager=pass_manager,
+            main_scope=MAIN_SCOPE,
+        )
+
+        assembled_program = assemble_starknet_contract(
+            preprocessed,
+            main_scope=MAIN_SCOPE,
+            add_debug_info=False,
+            file_contents_for_debug_info=file_contents_for_debug_info,
+        )
+
+        return json.dumps(
+            assembled_program.Schema().dump(assembled_program),
+            indent=4,
+            sort_keys=True,
+        )
