+++ starknet_py/compile/map.cairo
