+++ starknet_py/compile/mock-contracts/base.cairo
