+++ starknet_py/compile/mock-contracts/inner/inner.cairo
